requirements.txt
ccxt
requests
websocket-client
matplotlib
customtkinter
